Refactor events.ts into api.ts and atoms.ts
