+++ events_frontend/src/components/SelectSomething.tsx
@@ -1,5 +1,7 @@
 import {
+  Box,
   Button,
+  HStack,
   Menu,
   MenuButton,
   MenuItemOption,
