+++ events_frontend/src/data/atoms.ts
@@ -2,63 +2,18 @@ import { atom } from "jotai";
 import { groupBy, uniqBy } from "lodash";
 import { FilterOption } from "../components/SelectSomething";
 import { EVENT_TYPE_COLORS, WHENEVER } from "./constants";
-import { formatDate, formatEventTime, formatRelativeTime } from "./dateFormats";
 import { DateRange, getDateRange } from "./dateRanges";
 import { atomWithReset } from "jotai/utils";
-
-// Here's how the API returns info about the events. Some of these
-// like startTimeFormatted are derived when we load in the data to simplify
-// the EventCard display logic.
-export interface Event {
-  id: string;
-  name: string;
-  date: string; // 8601 formatted date string, pacific time assumed
-  type: string;
-  start: EventTime | null;
-  end: EventTime | null;
-  location: string;
-  address: string;
-  description: string;
-  cost: string;
-  link: string | null;
-  dateFormatted: string;
-  // Derived fields:
-  // start and end times converted to a string for display
-  startTimeFormatted?: string;
-  endTimeFormatted?: string;
-  // "3 days from now" or "1 hour ago"
-  relativeStartTime?: string;
-}
-// This is how we store events for each day
-export type EventsByDate = Record<string, Event[]>;
-// The API represents event times like this:
-export interface EventTime {
-  hour: number;
-  minute: number;
-  second: number;
-}
-
-// TODO: Move this to it's own file
-export const fetchEvents = async () => {
-  const url = "/example.json";
-  const json = await fetch(url).then((res) => res.json());
-  return json.events.map((event: Event) => {
-    return {
-      ...event,
-      dateFormatted: formatDate(event.date),
-      startTimeFormatted: formatEventTime(event.start),
-      endTimeFormatted: formatEventTime(event.end),
-      relativeStartTime: formatRelativeTime(event.date, event.start),
-    };
-  });
-};
+import { Event } from "./api";
 
 // A note about Jotai Atoms:
+//
 // Atoms make it super easy to work with derived state. In this app, most
 // of the atoms are derived from the eventsAtom below. When it changes,
-// the rest of them change automatically if needed. It makes derived state
+// the rest of them change automatically if needed. It makes state
 // management a breeze. If you're used to Redux, this has the same benefits
 // plus computed properties but with almost no boilerplate.
+// https://jotai.org/
 
 // Here's where all the events are stored once fetchEvents runs.
 export const eventsAtom = atom<Event[]>([]);
